refactor: 边框7 tsx 改造
