+++ packages/datav-vue3/components/BorderBox7/src/index.less
@@ -1,0 +1,32 @@
+.dv-border-box-7 {
+  position: relative;
+  width: 100%;
+  height: 100%;
+
+  .dv-border-svg-container {
+    position: absolute;
+    top: 0px;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+
+    & > polyline {
+      fill: none;
+      stroke-linecap: round;
+    }
+  }
+
+  .dv-bb7-line-width-2 {
+    stroke-width: 2;
+  }
+
+  .dv-bb7-line-width-5 {
+    stroke-width: 5;
+  }
+
+  .border-box-content {
+    position: relative;
+    width: 100%;
+    height: 100%;
+  }
+}
