+++ packages/docs/docs/Border/BorderBox7/demo.vue
@@ -1,9 +1,13 @@
 <template>
   <div demo-bg>
-    <dv-border-box-7>
+    <dv-border-box7>
       <div dv-bg>
         dv-border-box7
       </div>
-    </dv-border-box-7>
+    </dv-border-box7>
   </div>
 </template>
+
+<script setup lang="ts">
+import { BorderBox7 as DvBorderBox7 } from '@kjgl77/datav-vue3'
+</script>
